requirements.txt
# Smart Digital Food Menu Database Requirements
# Install with: pip install -r requirements.txt

# Core API Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database Connectors
# SQLite support is included with Python
mysql-connector-python==8.1.0

# Additional useful packages for development
python-dateutil==2.8.2
requests==2.31.0

# Development tools (optional)
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Production deployment (optional)
gunicorn==21.2.0
python-dotenv==1.0.0